requirements.txt
Django==4.0
djangorestframework==3.14.0
django-filter==23.2
drf-yasg==1.21.7
python-dateutil==2.8.2
channels==4.0.0
requests==2.31.0
drf_extra_fields==3.7.0
django-imagekit==4.1.0
pillow==10.0.0
channels_redis==4.1.0
daphne==4.0.0